+++ frontend/src/app/components/admin/admin-absences/admin-absences.component.ts
@@ -12,6 +12,7 @@ import { SessionService } from 'src/app/services/session.service';
 import * as cloneDeep from "lodash/cloneDeep";
 import { MatPaginator, PageEvent } from '@angular/material';
 import { ExcelService } from 'src/app/services/excel.service';
+import { DateFormingService } from 'src/app/services/date-forming.service';
 
 @Component({
   selector: "app-admin-absences",
@@ -54,14 +55,6 @@ export class AdminAbsencesComponent implements OnInit, OnDestroy {
   private stringFilter = new Subject();
   public stringFilter$ = this.stringFilter.asObservable();
 
-  private getDate(date) {
-    let formatedDate = [];
-    for (let i = 0; i < 3; i++) {
-      formatedDate.push(date[i]);
-    }
-    return formatedDate;
-  }
-
   constructor(
     public api: ApiCommunicationService,
     private route: ActivatedRoute,
@@ -69,7 +62,9 @@ export class AdminAbsencesComponent implements OnInit, OnDestroy {
     private router: Router,
     private session: SessionService,
     private excelService: ExcelService,
+    private dateFormingService: DateFormingService
   ) {}
+
 
   ngOnInit() {
     this.route.data
