+++ frontend/src/app/services/date-forming.service.ts
@@ -1,0 +1,15 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class DateFormingService {
+
+  constructor() { }
+
+  private getDate(date) {
+    let formatedDate = [];
+    for (let i = 0; i < 3; i++) {
+      formatedDate.push(date[i]);
+    }
+    return formatedDate;
+  }
+}
